requirements.txt
accelerate>=1.6.0
diffusers==0.33.1
einops==0.8.1
einops-exts==0.0.4
facexlib==0.3.0
fastapi==0.115.12
fsspec==2025.3.2
gradio>=5.31.0
# gradio_imageslider==0.0.20
httpx==0.28.1
k-diffusion==0.1.1.post1
kornia==0.8.0
Markdown==3.8
# matplotlib==3.10.1
ninja==1.11.1.4
numpy==1.26.4
omegaconf==2.3.0
open_clip_torch==2.32.0
openai-clip==1.0.1
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.2.1
pytorch-lightning==2.5.1.post0
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
timm==1.0.15
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
triton>=3.2.0
urllib3==2.4.0
uvicorn==0.34.2
wandb==0.19.10
webdataset==0.2.111
# xformers==0.0.32
xxhash
psutil>=5.9.0
# runpod>=1.7.0

